:recycle: Extract routes.go from pipeline_handler.go
